Protype of file ref counting with N:N mapping
If there is an n:n mapping between bundles and files, we should
add ref-counting for file->bundle references.

For instance, it is not safe to delete a file if the file references
a bundle.

File references are stores as empty files in s3 with a name:
	files-by-bundle/{bundle_fqid}{file_fqid}

Find refs with
	files-by-bundle/{bundle_uuid}*

Reverse ref lookups will be more difficult, unfortunately.